\encoding{UTF-8}
\name{indicspecies-package}
\alias{indicspecies-package}
\alias{indicspecies}
\docType{package}
\title{
 Functions to assess the strength and significance of relationship of species site group associations
}
\description{
This package provides a set of functions to assess the strength of species site group associations. It is also possible to check the statistical significance of such associations.
}
\details{
\tabular{ll}{
Package: \tab indicspecies\cr
Type: \tab Package\cr
Version: \tab 1.6.5\cr
Date: \tab 2008-10-23\cr
Depends: \tab permute \cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
Use \code{\link{strassoc}} for assessing strength of associations and \code{\link{signassoc}} to assess the statistical significance of the associations. A much more complete function is \code{\link{multipatt}}.
}
\author{
Miquel De Caceres, Florian Jansen

Maintainer: Miquel De Caceres <miquelcaceres@gmail.com>
}
\references{
De \enc{Cáceres}{Caceres}, M. and Legendre, P. 2009. Associations between species and groups of sites: indices and statistical inference. Ecology 90(12): 3566-3574.

De \enc{Cáceres}{Caceres}, M., Legendre, P., Moretti, M. 2010. Improving indicator species analysis by combining groups of sites. Oikos 119(10): 1674-1684.
}
\keyword{ package }