\name{wetland}
\alias{wetland}
\docType{data}
\title{Wetland vegetation data set}
\description{
  Vegetation of the Adelaide river alluvial plain (Australia). This data set was published by Bowman & Wilson (1987) and used in Dale (1988) to compare fuzzy classification approaches.
}
\usage{data(wetland)}
\format{
  A data frame with 41 sites (rows) and 33 species (columns). Abundance values are represented in abundance classes.
}
\source{
Bowman, D. M. J. S. and B. A. Wilson. 1986. Wetland vegetation pattern on the Adelaide River flood plain, Northern Territory, Australia. Proceedings of the Royal Society of Queensland 97:69-77.}
\references{
Dale, M. B. 1988. Some fuzzy approaches to phytosociology. Ideals and instances. Folia geobotanica et phytotaxonomica 23:239-274.}
\examples{
data(wetland)
}
\keyword{datasets}
